backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-ppl-pdf"
version = "1.0.0"
description = "Sistema RAG para consultas sobre documentos PDF"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.34.0",
    "chromadb>=0.4.22",
    "pypdf>=3.17.0",
    "pdfplumber>=0.10.0",
    "sentence-transformers>=2.2.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
]
